+++ .idea/misc.xml
@@ -8,7 +8,6 @@
       <list>
         <option value="$PROJECT_DIR$/springBasic/pom.xml" />
         <option value="$PROJECT_DIR$/HomeWork3/pom.xml" />
-        <option value="$PROJECT_DIR$/HomeWork1/pom.xml" />
       </list>
     </option>
   </component>
