+++ springBasic/src/main/resources/static/index.html
@@ -1,0 +1,73 @@
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>Page</title>
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
+
+    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.8.0/angular.min.js"></script>
+    <script src="index.js"></script>
+</head>
+<body ng-app="front-app">
+<div class="container" ng-controller="indexController">
+    <h1>EShop</h1>
+    <h5>{{product}}</h5>
+<!--    <table class="table">-->
+<!--        <thead>-->
+<!--        <tr>-->
+<!--            <td>ID</td>-->
+<!--            <td>Title</td>-->
+<!--            <td>Price</td>-->
+<!--            <td>Category</td>-->
+<!--        </tr>-->
+<!--        </thead>-->
+<!--        <tbody>-->
+<!--        <tr n>-->
+<!--            <td>{{product.id}}</td>-->
+<!--            <td>{{product.title}}</td>-->
+<!--            <td>{{product.price}}</td>-->
+<!--            <td>{{product.category}}</td>-->
+<!--        </tr>-->
+<!--        </tbody>-->
+<!--    </table>-->
+
+    <table class="table">
+        <thead>
+        <tr>
+            <td>id</td>
+            <td>title</td>
+            <td>price</td>
+            <td>Category</td>
+            <td></td>
+            <td></td>
+        </tr>
+        </thead>
+        <tbody>
+        <tr ng-repeat="p in productsPage.content">
+            <td>{{p.id}}</td>
+            <td>{{p.title}}</td>
+            <td>{{p.price}}</td>
+            <td>{{p.category}}</td>
+            <td>
+                <button class="btn btn-primary" ng-click="showInfo(p)">info</button>
+            </td>
+            <td>
+                <button class="btn btn-danger" ng-click="delete(p)">X</button>
+            </td>
+        </tr>
+        </tbody>
+    </table>
+    <label>{{pageNumber}}</label>
+    <br>
+    <label>total pages: {{pages}}</label>
+    <br>
+
+    <div class="btn-group" role="group" aria-label="группа кнопок" th:each="pIndex : ${totalPages}">
+        <a class="btn btn-primary" text="1" ng-click="loadProducts(1)">1</a>
+        <a class="btn btn-primary" text="2" ng-click="loadProducts(2)">2</a>
+    </div>
+
+</div>
+</body>
+</html>
