--- springBasic/src/main/resources/static/page.html
@@ -1,55 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Page</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.8.0/angular.min.js"></script>
-    <script src="index.js"></script>
-</head>
-<body ng-app="front-app">
-<div class="container" ng-controller="indexController">
-    <h1>EShop</h1>
-    <h5>{{product}}</h5>
-    <table class="table">
-        <thead>
-        <tr>
-            <td>ID</td>
-            <td>Title</td>
-            <td>Cost</td>
-        </tr>
-        </thead>
-        <tbody>
-        <tr>
-            <td>{{product.id}}</td>
-            <td>{{product.title}}</td>
-            <td>{{product.cost}}</td>
-        </tr>
-        </tbody>
-    </table>
-<!--    <table class="table">-->
-<!--        <thead>-->
-<!--        <tr>-->
-<!--            <td>id</td>-->
-<!--            <td>title</td>-->
-<!--            <td>price</td>-->
-<!--            <td>actions</td>-->
-<!--        </tr>-->
-<!--        </thead>-->
-<!--        <tbody>-->
-<!--        <tr ng-repeat="p in productsPage.content">-->
-<!--            <td>{{p.id}}</td>-->
-<!--            <td>{{p.title}}</td>-->
-<!--            <td>{{p.price}}</td>-->
-<!--            <td>-->
-<!--                <button class="btn btn-primary" ng-click="showInfo(p)">info</button>-->
-<!--            </td>-->
-<!--        </tr>-->
-<!--        </tbody>-->
-<!--    </table>-->
-<!--    <button class="btn btn-primary" ng-click="loadProducts()">refresh</button>-->
-</div>
-</body>
-</html>
