--- springBasic/src/main/resources/templates/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.8.0/angular.min.js"></script>
-    <script src="../static/index.js"></script>
-</head>
-
-<body>
-    <div class="container">
-        <div class="mb-3">
-<!--            <a th:href="@{/users/show_all}" class="btn btn-primary">List of users</a>-->
-            <a th:href="@{/products/show_all}" class="btn btn-primary">List of products</a>
-        </div>
-
-    </div>
-</body>
-</html>
