--- springBasic/src/main/resources/templates/users/products.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Products</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-
-</head>
-<body>
-    <div class="container">
-        <h1 th:text="${user.name} + '`s products' "/>
-        <br>
-        <table  class="table table-striped">
-            <tr>
-                <td>ID</td>
-                <td>Title</td>
-                <td></td>
-            </tr>
-            <tr th:each="p : ${products}">
-                <td th:text="${p.id}"></td>
-                <td th:text="${p.title}"></td>
-                <td>
-                    <a th:href="@{'/products/' + ${p.id}}" class="btn btn-primary">Info</a>
-                </td>
-            </tr>
-        </table>
-        <span th:text="'Всего продуктов: ' + ${products.size()}"/>
-        <div class="p-3 col-6">
-
-            <a th:href="@{/index}" class="btn btn-primary">index</a>
-        </div>
-
-    </div>
-</body>
-</html>
