--- springBasic/src/main/resources/templates/users/user.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>User</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-</head>
-<body>
-<div class="container">
-    <h1>User information</h1>
-    <br>
-    <table  class="table table-striped">
-        <tr>
-            <td>ID</td>
-            <td>Name</td>
-        </tr>
-        <tr>
-            <td th:text="${user.id}">id</td>
-            <td th:text="${user.name}">name</td>
-        </tr>
-
-    </table>
-     <div class="mb-3">
-         <a th:href="@{/users/show_all}" class="btn btn-primary">List of users</a>
-     </div>
-
-
-</div>
-</body>
-</html>
